test_parser/pyproject.toml
[tool.poetry]
name = "test-parser"
version = "0.1.0"
description = ""
authors = ["Erik Wrede <erikwrede2@gmail.com>"]
readme = "README.md"
packages = [{include = "test_parser"}]

[tool.poetry.dependencies]
python = "^3.10"

rustberry = { file = "../target/wheels/rustberry-0.6.0-cp310-cp310-macosx_10_7_x86_64.whl" }
graphql-core = "^3.2.3"
strawberry-graphql = "^0.151.2"
pyinstrument = "^4.4.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
